Fixed regex
The regex in url_bases in moto.s3.urls was too broad. Bucket names cannot be zero characters, and the dot separating the bucket name from the rest of the domain name is not optional. This overly-broad regex caused URLs that should have gone to the moto.s3bucket_path backend go to the moto.s3 backend instead.