AWSServiceSpec: Fix `TypeError` exceptions within json.load()
The load() method provided by the built-in JSON module does not accept a
byte-type value in Python 3.5 (or versions before), and will raise an
exception if one is passed.

For details, please see: https://bugs.python.org/issue17909

Thus, for better compatibility, we'd better decode the content of the
JSON file before passing it to the parser, instead of letting the module
to guess the encoding.