Support optional Source, parse from header
The Email ``from`` header is either formatted as ``name <address>`` or ``address``.

This commit will use `parseaddr` to extract a ``(name, address)`` tuple, which we will use the ``address`` to check if it's verified.

Also support the case where ``Source`` is omitted (which AWS requires the ``from`` header to be set).