Rework LaunchTemplateBackend to be keep only one copy of a template, and be ordered
The original LaunchTemplateBackend kept two copies of a template, one
for referencing it by name and one for referencing it by id. This change
switches to using one copy, by id, and adding a lookup dict for mapping
names to ids.

Additionally, to fix the python2 test ordering issues, the launch
template dict was changed to an OrderedDict.