Extend test cert validity to 2049
to make tests pass in 2027

Background:
As part of my work on reproducible builds for openSUSE, I check
that software still gives identical build results in the future.
The usual offset is +15 years, because that is how long I expect
some software will be used in some places.
This showed up failing tests in our package build.

See https://reproducible-builds.org/ for why this matters.

Note: I tested that i586 builds still pass tests on 2036-12-01
so are not affected by the year 2038 bug yet.

i586 builds in 2037 suffer from a y2038 bug (unrelated to this change):
 botocore/utils.py parse_timestamp threw a
  ValueError: Invalid timestamp "2174774434": Parser must be a string or character stream, not long