Don't resurrect exported resources.
stack.exports has the side effect that, if a resource
in the exports resource map does not exist, it gets
created. Incidentally when this function is called, most
resources don't exist, so anything created will be
instantly resurrected.

The solution is simple: just get the export names before
deleting the stack.