S3: Implement CORS headers in OPTIONS requests
The default headers from CORS plugin was not enough to mock S3 CORS Rules behavior
We need to manually define response headers depending on the bucket rules

- Fixed a typo `ExposedHeader` --> `ExposeHeader` (defined as such in AWS documentation)
- I implemented OPTIONS method instead of relying on the defaults from Flask
- I removed a line that looked like a bug:
```python
self.headers = request.headers
``` 
This was making our current headers point to the Request header, which is not mutable.

- Iterated through CORS rule to return some CORS headers
This is not a perfectly accurate strategy: I didn't check but according to the docs, when multiple CORS Rules are defined, S3 returns the most relevant (I suppose the most permissive one, depending on the request)
In the strategy I implemented, the last rule takes precedence over the other rules.
It's not ideal but I consider that defining multiple rules is an edge case.