add back OPTIONS key response after implementing CORS in #4497
In #4497 I forgot to implement OPTIONS on key response, I only added unit tests for bucket responses (queries on `/`)

The OPTIONS key response on a real S3 endpoint is essentially the same as the bucket response.
If the bucket doesn't exist, return 403. If the bucket exists, return 200 with the bucket CORS headers, disregarding if the key exists or not.
So I used the same method for both key and bucket responses.