Added support for custom AWS Lambda images.
Currently the image used for executing Lambda functions is hardcoded:

https://github.com/spulec/moto/blob/e84cc20abe5b92e3bb77bcddd9240d4233a9524e/moto/awslambda/models.py#L613

Unfortunately it seems that development has also stalled, the last
release at the time of writing being over one year ago:

- https://github.com/lambci/docker-lambda
- https://hub.docker.com/r/lambci/lambda/tags

Because of this, there is no image built for Python 3.9 which means it
is not possible to use this with moto.

With this change it is possible to switch to using a supported fork:

- https://github.com/mLupine/docker-lambda
- https://hub.docker.com/r/mlupin/docker-lambda/tags

This is achieved by setting the `MOTO_DOCKER_LAMBDA_IMAGE` environment
variable to override this as required.