fix(ec2): handle non-existent template name for describe-launch-templates
Currently, when calling `ec2 describe-launch-templates` with a template name
that does not (yet) exist, an unhandled exception occurs and HTTP 500 is
returned.

This commit adds a check for the existence of the passed launch template name.
If a template by this name does not exist, an `InvalidLaunchTemplateNameNotFoundError` is raised.
It also renames the current `InvalidLaunchTemplateNameError` to `InvalidLaunchTemplateNameAlreadyExistsError`,
so that both cases are representable.