S3: Fix deadlock in list_object_versions after multipart upload
After a completed multipart upload, the final FakeKey object contains a
reference to the FakeMultipart object, which in turn contains references
to the FakeKey objects used to represent each uploaded part. The
FakeMultipart object, and its FakeKey objects, will have been disposed
when the uploaded completed.

In list_object_versions, meanwhile, all FakeKey objects are deep copied,
which results in copying and reading the values of the disposed FakeKey
objects referenced by FakeMultipart objects.

Ultimately, this causes an exception in FakeKey.value(), since it's
trying to read the value from a file that was closed when it was
disposed. The exception is raised with a lock acquired, which results
in a deadlock the next time the value is read.