Remove dependency on pytz
This removes the dependency on pytz, which was used to provide the UTC
offset information for timezone aware datetime objects. While that was
helpful for Python versions up to 3.1, UTC offset information is
included in the standard library since Python 3.2. As moto doesn't
support Python versions <3.6 anymore, using pytz isn't necessary anymore
either.