typing for backends
This allows typing to work properly when starting with a BackendDict object.
Most functions were ending up as Any before, but now mypy can figure out what they are

This turned up a lot of mypy errors, which I mostly didn't try to resolve.
A lot of them are in create_from_cloudformation_json() functions.