tighten up typing for get_backend()
It's unfortunately a lot of boilerplate, but this allows
users of get_backend() to end up with a correctly typed backend.

I think we'd need to separate the typing into stub files, and then
write stub generation code (like boto3-stubs uses) if we wanted
to get the same effect without the boilerplate.

One bug fix: rdsdata is named 'rds-data' in boto3.