RDS: Default value for `DBCluster.CopyTagsToSnapshot` is `False`
From the [AWS documentation](https://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_CreateDBCluster.html):
"Specifies whether to copy all tags from the DB cluster to snapshots of the DB cluster. The default is not to copy them."

Interestingly, this default behavior was already covered by a test[^1], but it was only succeeding because we weren't
properly serializing the boolean value.  This commit also fixes that.

Another test, `test_add_tags_to_cluster_snapshot`, was relying on the incorrect default value being set to `True`,
so that test has been updated to explicitly request that tags be copied from the cluster to any snapshots.

[^1]: https://github.com/getmoto/moto/blob/d11b181e19b2e95255f2f663e84c4e46e20612cc/tests/test_rds/test_rds_clusters.py#L211