Consolidate `Neptune` and `RDS` backends (#7925)
Despite being documented as two distinct services (with two distinct boto3 clients), the Neptune and RDS services share a single AWS backend (RDS). This is reflected in the botocore service definition for Neptune, which lists `rds` as its endpoint. Calling the `DescribeDBClusters` API method with either the `neptune` or `rds` boto3 client results in the exact same list of clusters--no filtering is done on the AWS backend. Neptune clusters are commingled with Aurora clusters.