Validate RDS snapshot identifiers
Snapshot identifiers need validation that's very similar to that of instance
identifiers, just the maximum length is 255 instead of 63.

AWS actually produces three different kinds of error messages for invalid
snapshot identifiers, depending on what it does not like.
We need to keep track of the operation that caused the error, because the
error message needs to include details about the parameter name.