RDS: Update DBCluster MultiAZ Property
The documentation clearly states that a DBCluster is considered multi-az
if it has DBInstances in multiple availability zones.  I'm not sure if
the existing check here was correct, but it is covered by tests that I
don't want to break, so I'm just adding a new conditional for determining
multi-az status.